Move to treating `backend` as a Python package
In the process of implementing functional testing of the service and
database layer of the backend, it became simpler to treat the `backend`
directory as the top-level Python package, adding `__init__.py` files
around. This led to moving toward relative imports in all project
locations, though the imports could be absolutely from `backend...`.

This has an impact of needing to start uvicorn from the workspace and
fixing the relative path of static assets.

In this commit, we see the basis of a simple unit test with fixtures
to setup and teardown a full-on database with tables in Postgres, but
we do not have mock data or actual assertions. That will come next.

Still chipping away at #3