Establish simple PermissionService access control
Moving to a model where Users and Roles can have Permissions associated
with them. However, for the time being, the emphasis is on Roles. User
specific permissions should work, but need additional testing and are
more likely to be handled by feature-specific logic.

Next steps are surfacing this in an admin panel for managing roles,
users within roles, and permissions of roles from the front-end and API.

Continued progress on #3.