Convey user's permissions in profile API for UI
Removing circular dependency of Permission <-> User (and Role) by making
a Permission a child of User/Role. Led to some slight refactoring of
`grant` in PermissionService.

With Permissions being conveyed to the front-end, the UI should be able
to perform some basic tests about what to surface using a similar logic
as the backend for testing access.

More progress on #3