Check permissions client-side with match logic
We scan permissions using the same logic as on the service side service.
This is done in an observable pipeline, which seems like the right call,
but it will be interesting to see how this impacts downstream code. For
the most part, permission checking on the client side is for UI surface
decisions only and embedded in the templates.