Add access control system with admin management
In the process of implementing functional testing of the service and
database layer of the backend, it became simpler to treat the `backend`
directory as the top-level Python package, adding `__init__.py` files
around. This led to moving toward relative imports in all project
locations, though the imports could be absolutely from `backend...`.

This has an impact of needing to start uvicorn from the workspace and
fixing the relative path of static assets.

Introducing the basis of a simple unit test with fixtures
to setup and teardown a full-on database with tables in Postgres, but
we do not have mock data or actual assertions. That will come next.

The model has Users and Roles with Permissions associated to them. 
However, for the time being, the emphasis is on Roles. User
specific permissions should work, but need additional testing and are
more likely to be handled by feature-specific logic.

We scan permissions using the same logic as on the service side service.
This is done in an observable pipeline, which seems like the right call,
but it will be interesting to see how this impacts downstream code. For
the most part, permission checking on the client side is for UI surface
decisions only and embedded in the templates.

Improved database reset script to reset sequence counts. This helps ensure
when new users register we don't have a conflicting id with the mock data.

Admin panel now has the ability to:

* List users
* List roles
* View individual Role details
  * Add permissions
  * Revoke permissions
  * Add members
  * Remove members