Log user out who authenticates with bad token
If a token is no longer valid, or does not validate with the key on
the server, we are responding with a 401 HTTP status code. In the
front-end client, we are intercepting 401 errors and logging
the user out of the front-end.

This raises an important secondary issue, in other places we may need to
fix response code to 403 if a user is validated with system but does
not have access to a resource.