Backend (#34)
* Create organizaton API routes

* Create organization service

* Create organization entity

* Create organization model

* Create method that gets organization by name

* Edit create method to remove update functionality

* Edit organization service create method's error message for duplicate

* Create registration api routes

This commit creates a file named 'registration.py' in the api folder,
which defines api routes used to update the Registrations table. These
routes will be used to connect the front end with the backend database.

This file defines the HTTP requests/methods to create, get, update, and
delete registrations from the Registrations table based on provided
conditions.

This commit also updates the main routes file to include the routes
created for registrations.

* Create registration service

This commit creates a file named 'registration.py' in the services
folder, which defines the RegistrationService used to update the
RegistrationEntity.

The RegistrationService defines methods to create, get, update, and
delete registrations based on various conditions.

* Create registration model

This commit creates a file named 'registration.py' in the models folder,
which defines the Registration model that serves as the data object for
a User's registration to an event.

The state of a Registration model is defined by the user id (which is
from the User table), the event id (which is from the Event table), and
an event status.

* Create registration entity

This commit creates a file named 'registration_entity.py', which defines
the schema for the SQL-alchemy table object for the RegistrationEntity.
The RegistrationEntity is used to relate Users and Events in order to
keep track of User's registrations and attendance for Events.

The RegistrationEntity defines columns for a unique registration id, the
user id (which is from the User table), the event id (which is from the
Event table), and an event status. It includes methods for converting a
RegistrationEntity to a Registration model and vice versa.

* Include registration routes in main routes file

This commit adds code to main.py to include registration routes. I
initially thought I completed this in a previous commit (e1150a1), but
I actually hadn't done it yet when I committed.

* Create event entity

* Create event model

* Create event service

* Create event API routes

* Update the definition of registration model.

This commit edited the Registration model to have the unique id (based
on the RegistrationEntity table) and allow it to be None. The goal for
doing this is to hopefully allow for the creation of Registrations in
the frontend without specifying the id, since that is autogenerated
by the database and not something we would know when creating a new
registration.

I also modified the registration service delete_registration() method
to directly use the registration id rather than querying based on
the user and event ids, since we should know the registration id when we
need to delete.

As I am writing this commit message, I realized there are further
updates to be made to the files associated with the Registration. I
don't want to forget, so here are two of the things I thought of:
1. Make the Registration id get update properly after registration is
created.
2. Update the Registration entity to/from model methods to reinclude
the id property.

* Update entity methods to include id.

This commit updates the entity methods to reinclude the primary key id.
It previously was not in there because I wasn't sure how to handle not
knowing what the id is until after creation, but I decided to allow
the Registration model id to be None to account for that brief period
of time where a registration hasn't been put into the table yet and
therefore doesn't have an id.

Further edits and implementation might need to be added to make this new
definition of the Registration model functional for handling
interactions between the frontend and backend.

* Add event api routes to main.py

* Separate create and update functionality

* Fix organization API routes

* Fix event API routes

* Fix default organization API routes

* Add user api routes and service methods

* Fix file imports and API routes

This commit fixes the imports between files. Before, I only had
"..entities" and so on for my imports, but I needed to write out the
full path name for it to actually get imported.

This commit also fixes the API routes. I previously used my original
format with just "app", and I had no prefix. I adjusted the API routes
instead to have prefixes as the other files do.

* Fix typeerror in user service

In user service, there is a function named "list," which caused a typeerror when I tried to use list[User], because it thought I was calling a function. For now, I moved my function where I return a list of users above the "list" function; in the future I think we should rename the function, but I am leaving it for now since this was part of the code we were given.

* Update user delete route to follow convention

* Fix API routes and file imports

* Fix type errors, duplicate import, and get by name method

* Fix relative imports

* Fix event entity in reset database

* Add documentation to list function

* Add Role Table Functionality (#26)

* Add Role Table Functionality

Added the Organization Role tables based on the role table in the other repository.

* Fix Semantics Issues in PR

This commit fixes semantic issues in the PR that Audrey mentioned.

* Fix Critical Merge Conflict Issue on Event Entity

Fixed an issue where existing merge conflict markup remained after the merge conflict was resolved.

In addition, this commit fixes import statements using relative paths that caused critical errors upon running with `honcho start`.

* Add organization dev data to reset database script

* Add mock data for Event, Registration, and Org Role tables to be added using reset database

* Add API tags for easier testing

* Remove link from CADS long description

* Change events dev data past/future testing

* Fix event datetime error in reset database script

* Add One-To-Many Database Relationships (#33)

* Create Relationship Between Event and Organization Table (#29)

This commit establishes the one-to-many relationship between the event and organization tables, and provides documentation for this connection.

This commit also modifies the existing entities and models to store the fields that establish this connection.

**`EventEntity`**
- Modifies `org_id` to become a mapped column with the foreign key of "id" from the "organization" table
- Creates a new `organization` field mapped to the "organization" table that stores the organization based on the `org_id`. This field also back-populates the new "events" field in the  `OrganizationEntity`.

**`OrganizationEntity`**
- Creates a new `events` field mapped to the "event" table that stores all of the events hosted by the organization. This field also back-populates the new "organization" field in the  `EventEntity`.

** Models**
- Modified both the `Event` and `Organization` models to specify the new fields from the modified entities.

* Fix Critical Errors with Relationships (#32)

This commit fixes critical errors with the existing one-to-many relationships.
- Solves circular import errors
- Add delete database script to eliminate cascading issues
- call `update_forward_refs` to prevent errors with missing classes

---------

Co-authored-by: jadekeegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: tdianaa <rtoneyd@gmail.com>
Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>
Co-authored-by: atoneyd <31745478+atoneyd@users.noreply.github.com>