Update Organization Page Frontend (#45)
* Modifications

* Create Organizations Component

Created the Organizations component in the `frontend` directory and hooked up the new component to routing. The new component will live at the route `/organizations`.

Also added the organizations page to the sidebar, and installed a few extra Material design components.

* Begin Designing Prototype Organizations Page

This commit explores potential designs for the organization page, weaving through the terror that is Angular Material.

* Convert Organizations Wireframe to Flex Boxes

This commit converts some of the layout Material elements to vanilla HTML elements using flex boxes and Tailwind style classes.

* Fix organization page styling

This commit fixes the organization page styling. It removes tailwind
inline styles and moves all styling to the organizations.component.css
file instead for clarity. It also fixes the alignment of items on the
organization page.

* Fix search bar color and card spacing

This commit changes the search bar text color from pink to blue-grey to
improve readability. It also adds a row-gap between the organization
cards.

* Add search bar styling

This commit styles the search bar and updates the spacing on the
organization cards to match the width of the search bar.

* Implement organization frontend features

This commit adds improved functionality to the organizations page. This
includes filtering events displayed based on search bar input with a
filter pipe and setting up the organization service. Some minor styling
was also added.

* Uninstall node and jest

This commit uninstalls node and jest as they should not be needed.

* Add dependencies to package lock

This commit adds other dependencies to the package lock.

* Fix icon colors

This commit fixes the organization card icon colors. For some reason,
they turned dark blue instead of white, so I had to add the class to the
css file and fix the color/fill.

* Connect organization backend with frontend page

This commit connects the organization backend table with the frontend
organization page. This allows for the frontend to pull from the
organization table and use the organization development data created.

* Refactor html to use angular mat ui

This commit updates the HTML to use Angular mat ui. We previously used
divs because the Angular mat ui styles were troublesome to work with,
since they override other styles.

* Fix organization card styling

This commit fixes the order of the icon and the title in the mat card.
It also fixes the z-index of the details button.

* Edit Procfile to allow frontend to autoreload

This commit edits the procfile so that honcho start runs "ng serve
--poll=2000", which causes the frontend server to automatically reload.
Previously, the frontend server would not autoreload.

* Fix organization card icons

This commit fixes the links so that icons for an organization's other
accounts/sites don't appear unless a link actually exists (link not
empty). It also restricts the number of icons shown to be three to avoid
overflow.

* Change organization page view permissions

This commit allows any user to view the Organizations Page without
being signed in and authenticated. It removes the canActivate condition
as well as adds a link to the Organizations Page to the sidebar.

* Delete comments organization component

This commit removes commented out variables that were used to test
the frontend before the dev data was created/merged in from backend.

* Fix small organization page issues

This commit fixes some issues with the Organization page pointed out
during CR. It makes the icon links open in a new tab rather than the
current one, removes extra characters from organization descriptions,
and fixes the organization logos so they aren't forced to be a
specific height.

* Change WiCS logo link

This commit changes link to the WiCS logo to improve the design.

* Clean up html and css

This commit fixes class names and deletes commented out html.

* Add organization page documentation

This commit adds documentation comments to the Organizations component
and its corresponding HTML/CSS files; the Organizations service; and the
search bar Filter Pipe.

* Add params/return documentation to methods

This commit adds documentation for parameters and return types to the
Organization Service and the Filter Pipe.

* Improve Organization search bar filter

This commit improves filtering based on input for the Organization page.
It adds other conditions to the filter so that organizations that
contain words matching the short and long descriptions also appear.

---------

Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>