Backend Dev Data and Relationship Fixes (#46)
* Modifications

* Fix dev data and org role/registration

This commit fixes the dev data so it does not include ids. It also fixes
the OrgRole and Registration get/create methods so they return the full
information. Previously, both didn't return the extra information
stored from back populating from other tables. This was fixed by
creating Organization/RegistrationSummary models, updating the api
calls, and updating the to_model() methods themselves.

* Fix create header in organization services

This commit fixes the parameter type of every service (Event,
Organization, Registration, OrgRole) to be the summary version. The
EventSummary was also updated to not include the organization, since
you will not have the populated organization itself when you create the
Event.

* Remove commented out Line

This commit removes a commented out line in backend/models/event.py. I
initially was unsure if removing this line would break the database, so
I just commented out. However, the code now works without the line, so
the comment is not necessary.