Profile Page (#49)
* Modifications

* Create profile page and set up routes

This commit creates the profile page component and sets up its routes.

* Create HTML for profile page

This commit creates the HTML structure for the profile page to give a
basis for the page's user interface and functionality. Styling on mat ui
components has not been completed. Implementation of HTTP requests
to get info has also not been implemented.

* Add styling to card grid

This commit styles the card grid as a whole.

* Create events page

* Style profile page

This commit styles the profile page with a grid to display the profile
and organizations on one side the events on the other.

* Update profile service to get events and orgs

This commit updates the profile service to include methods to get the
events and organizations based on the user. This also required defining
various export interfaces in the profile service.

This commit also has some minor styling for the profile grid.

* Add event filter and update page html/css

This commit adds the event filter pipe, which filters the events based
on whether they are before or after the current date. This commit also
makes edits to the profile page html and css.

* Add documentation to profile page html and css

This commit adds a few comments of documentation to the HTML and CSS
files for the profile page component.

* Finish styling profile page

This commit completes the profile page's styling. Most are minor
fixes. Media queries were added to change the display based on screen
width. Font size increased.

* Fix event detail info font size

* Fix event detail time font size

* Add documentation and complete minor fixes

This commit adds documentation to all profile page-related files. This
includes the Event Filter Pipe, Profile Page CSS, Profile Page HTML,
Profile Page Component, and Profile Service files. The organization
service documentation was updated to match the format of the profile
service documentation.

Also, in the profile service, the EventSummary still had a list of
events in the model rather than the associated Organization. This was
fixed accordingly.

* Abstract interfaces away from main components

This commit moves all exported interfaces in components/services to
the app/models.module.ts. There were some repeated exported interfaces,
and some of the files were just getting too crowded with exports. So,
we decided it would be better to just put them in their own separate
file.

* Make dynamic user avatar for profile

This commit replaces the hard-coded profile image with a user avatar
containing the letters of the user's first and last name.

* Remove unnecessary comment in profile page html

This commit removes commented out code from the profile page. The code
was from something I was trying to work on but decided against doing
right now since it's low priority.

* Implement cancel registration button

This commit creates and implements the "Cancel Registration" button
for user events. This allows users to unregister for events they have
not attended yet.

It also refactors the backend "delete_registration" method to delete by
registration id rather than registration object, since deleting by
registration object does not work.

* Fix typo in documentation

---------

Co-authored-by: briannata <brianna.hy.ta@gmail.com>