Fixes minor errors on stage (#71)
* Modifications

* Merge branch 'stage' of https://github.com/comp423-23s/final-project-team-a3 into stage

* Fix merge errors

This commit fixes errors that occurred because of improper merging.

* Fix minor errors with recent stage updates

This commit fixes minor errors with the recent stage updates. This
includes some CSS, HTML, and documentation for various pages.

* Fix Event/EventSummary models

* Fix registration edge case

When a user doesn't have their profile information saved, a user id is not created. We force them to create a profile before registering.

---------

Co-authored-by: jadekeegan <97476936+jadekeegan@users.noreply.github.com>