Update Event Page Registration Functionality (#82)
* Update event registration button

This commit changes the event registration button so that it does not
appear if the user is already registered for an event. It instead
replaces it with a delete button.

In order to do so, some methods were added to the Events Component and
the Events Service. Also, the register method was refactored. The code
used to check if the user was already registered for an event was moved
to a separate checkIsRegistered method in the Event Service so it could
be used to dynamically update the buttons. A corresponding method was
made in the component to call the service method.

* Add documentation and fix minor errors

This commit adds the checkIsRegister() message to an if statement to
condition to completely prevent users from registering for an event
twice. Technically, this should be possible due to the frontend setup.
However, just a precaution!

This commit also fixes some documentation in the HTML and services.