Implement Admin Organization Controls (#90)
* Set up admin orgs view and manager controls

This commit sets up the organization admin list view page, as well as
the individual organization detail admin page. It also adds
functionality on the organization detail admin page to add and remove
organization managers.

* Clean up code and add documentation

This commit cleans up the code for the admin organization controls and
adds documentation to the new methods and files. Some comments were left
in the addManager/removeManager methods as they need to be updated.
Currently, they completely add or completely remove managers. However,
they should instead just "promote/demote" users to/from manager.

* Update membership type rather than deleting roles

This commit fixes the admin page so that it updates the membership type
in a "promote/demote" manner rather than completely creating or
removing an organization role. This is because some managers could have
previously been members, so their role in the organization should just
be updated.

* Fix unstarring to distinguish execs from admins

This commit fixes unstarring organizations so you aren't able to unstar
if you are an executive OR an admin. It also does some refactoring to
to ensure that executives are separate from admins.

* Fix typos in snackbars and add permissions to data

This commit fixes the snackbars when attempting to star/unstar orgs.
The snackbar messages weren't updated to reflect executive vs. manager
even though the implementation was.

This commit also adds organization admin permissions.

* Delete commented out code

* Fix minor formatting and docs issues