Fix organization star and org editor button (#100)
This commit fixes the organization star so it can be removed on the
main organizations page. Initially, it was going into the if statement
for the filter since the org role still exists, so I added to the
condition that it shouldn't go there if the status is 0.

This commit also fixes the org editor button, which was somehow
removed during merging.