Add Delete and Create Organization functionality to Admin Page (#101)
* Set up delete and create org implementation

This commit sets up the functionality to delete/create organizations,
including the buttons and methods to do so. It also makes some changes
to the OrgEditor component to begin refactoring it so it can be used for
creation as well.

This commit also attempts to add cascade settings to fields that back
populate for organizations, but it did not work so further changes must
be made.

* testing

* Fix organization cascades

Cascade delete code was in the wrong place; it should go in the parent entity file, under the child parameters.

* Refactor admin org table to auto-update

* Remove print statements

* Fix admin organizations list reload

This commit fixes the admin organizations list update after deleting
an organization. In some cases, it wouldn't auto-update with the new
organizations list. The current fix is to reload the page after
navigating to it, but ideally we will fix this at some point (along
with all the other excessive refreshes).

This commit also adds some styling to the delete organization button.

---------

Co-authored-by: jadekeegan <97476936+jadekeegan@users.noreply.github.com>