Update Typescript According to Conventions (#103)
* Convert event editor ts to arrow functions

* Convert events ts files

Document functions and convert to arrow functions

* Convert eventsFilter ts files

convert Number to number, String to string, and convert to arrow functions

* Convert admin organization ts files

convert to arrow functions and add documentation to functions

* Update files according to TS conventions

This commit updates the modified files to use arrow functions.

* Fix ngOnInit and fix org editor

This commit reverts ngOnInit back into a normal method rather than
an arrow function, as it wasn't actually working as an arrow function.

This commit also fixes the org editor so it autopopulates for existing
events.

* Update Organization Editor title

This commit fixes the html for the organization editor so that the
organization editor has a "Create Organization" title when creating
an org and a "Update Organization" title when updating an org.

This commit also changes the name of the Organization Editor page to be
Organization Editor rather than "Update Organization".

* Fix title of event editor

This commit fixes the title of the event editor to be dynamic based
on whether a user is creating vs. updating an event. It also changes the
page name to be "Event Editor" rather than "Edit Event".

* Change String -> string for Org models

This commit changes String wrapper to string primitive for any model
associated with organizations.

* Fix docs and whitespace

* Fix line break

---------

Co-authored-by: briannata <brianna.hy.ta@gmail.com>