Protect API Routes (#109)
* Create tests for organizations backend

* Add authentication security for event api routes

* Add security to org role api/services

This commit adds a check to ensure that users cannot create/delete org
roles without the proper permissions in the backend. This ensures that
people cannot go to the docs page and create/delete org roles.

* Protect organization api routes

* Fix Org Editor Permissions

This commit fixes the org editor permissions. Previously, the "create"
version of the page would be visible to anyone who is an organization
manager. This commit ensures that it is only available to admins
specifically.

* Add security to mutating registration API routes

This commit adds security/permission checks to the mutating registration
API routes (create/update status/delete/clear all) in order to prevent
users from going into the docs and making changes to the registrations
table.

* Fix whitespace

* Fix tags and add comment

This commit fixes the tags in the api route files so they do not include
"Detail" as a suffix in order to improve the readability of the docs
page. It also adds a comment explaining the -1 org id in the org-editor
file, which is used to indicate that a new organization (which does not
yet have an id since it hasn't been added to the table) is being made.

* Fix Event API Tags

Edited Event tags from EventDetail back to Event.

---------

Co-authored-by: briannata <brianna.hy.ta@gmail.com>
Co-authored-by: tdianaa <rtoneyd@gmail.com>
Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>