Fix logic errors/bugs in services requiring auth
This commit fixes some logic errors and other bugs that were occurring
in files that were refactored to require authorization on mutating
API routes and their corresponding service methods. One important
change is in the Event, Organization, and Registration services. Other
services (OrgRole and/or Event) were injected in so model-specific
service methods could be used in them.