Add Public and Private Organization Functionality to Frontend
This commit deprecates the "starOrganization" functions in the organization services and has replaced this function with "toggleOrganizationMembership". If an organization is marked as public, any user is able to join and leave and organization. If an organization is not marked as public, users are unable to join the organization and receive a snackbar tooltip explaining why.

In addition, there is now a snackbar that appears for a user to confirm that they want to remove membership from an organization. This was added since, especially with private organizations, users would likely want to confirm they are leaving an organization if they press the button by mistake.

The star icons have also been replaced with more appropriate icons. These icons can be changed depending on what we think is best for our design!