Add Public and Private Event Functionality to Frontend
This commit edits the event filter pipes to show and hide organizations based on whether a user should be able to use them. If an organization's event is marked "public", the event will be shown on the events page (and org detail page). Users, regardless of membership to the org, would be able to register for the event. If an organization's event is marked "private", then only members of that organization are able to see and register for the event.

This commit also adds some extra test data to show this is working.

Also, This commit fixes a bug where the edit event button shows up regardless of admin status.