Make Coworking tests less sensitive to time
This commit moves times from globals to a fixture. This was necessary
because as our test suite grows the times calculated at the start
became heavily skewed relative to when individual tests began running.
The fixture recalculates time offsets commonly used at the start of
every individual test run as a fixture.

* Move toward relative times being calculated as a fixture
* Bring AvailabilityList test up to new time fixture standards
* Refactor Coworking OperatingHours Tests to use new time fixture
* Refactor Coworking Reservation Tests to use new time fixture
* Rename times module to time
* Update dev data reset script