Automatically transition/GC Reservations to states
A few of the state transitions for a reservation are time-sensitive.
For example, a user has five minutes after beginning a reservation draft
to commit to it before it is automatically cancelled. This commit adds
a first pass of functionality to realize this. It also improves some
documentation.

This method for GC is a bit of a hueristic and we likely need to
implement a cron-based job that periodically goes through and does
a full pass. As currently implemented, only reservations which come up
in searching for availability will get GC'ed. In practice, this should
mean that all meaningful reservations are GC'ed, from the user's perspective.
From the system's perspective, however, there is a likelihood
of some reservations falling through the crack. Most susceptible are
those not converted to "checked out" due to no future user searches
turning up those reservations as they were in the past. Setting up
an issue on GitHub to track this future improvement.