Refactor Events Components
This commit is a major refactor to the event components. With this new commit, the events-page and event-editor components are nested into one folder and share the same component. This is to match convention with the profile page, as well as the new coworking page - and will make navigating the frontend easier.