Refactored Organization Components
This commit is a major refactor to the organization components. With this new commit, the organizations-page, org-details, org-roster, and org-editor components are nested into one folder and share the same service.  This is to match convention with the profile page, events page, and coworking page - and will make navigating the frontend easier.