Add Organization Detail Component
This major commit adds the organization detail frontend component and all supporting widgets to the organizations module.

There is also an interesting thing with how this was set up. Since organization is now a standalone module, routing is also taken care of by the organization module. There were some difficulties with getting this set up, but the idea is that the new organization routing module starts managing routing at the `/organizations` level. So, the route for organization page would be "", since it is at the *root* of where the new routing module is managing. The detail pages would then be ":id", since their real route should be "organization/:id".