Improve Organization feature
* Update organization details page to use resolver

This commit refactors the Organization Details page to use a functional
resolver to inject the Organization into the application rather than
using subscriptions.

* Update organization details title resolver

This commit fixes the organiztion details page title resolver to ensure
that it does not make an unnecessary extra HTTP request to get the
organization information. This is done by making title/title resolver a
child so it can inherit the parent route data!

* Add Permission Guard to Organization Editor
* Refactor Organization into the Resolver
* Add Social Media Icon Service (#92)
* Add Social Media Icon Service
* Remove Spec File

* Update the admin organization page to use resolvers

This commit refactors the admin organization page to use route resolvers
for consistency across components. It unfortunately does not auto update
however as the route data does not get refreshed when the database
changes.

* Clean up imports

This commit removes unnecessary imports from the organization details
component and the organization editor component.

* Improve Organization Slugs (#90)
* Add shorthand to organization backend
* Add organization shorthand to frontend
* Add automatic slug generation for org creation
* Add documentation to change method
* Add alembic migration to add shorthand column (#97)
* Add form validation (#91)
* Add form validation
* Add hint specification
* Update Organization Details Page to use short description if no long description

---------

Co-authored-by: Brianna Ta <37271052+briannata@users.noreply.github.com>

* Clean Up Organization Backend (#95)

* Check admin permission for org update

* Make get from slug more idiomatic

* Fix organization create id nit

* Add more specific error handling

* Adjust nav toolball z index

* Implement bandaid fix for delete organization reload

This commit adds a filter to remove the organization from the array
so the HTML automatically updates the list data. Without this, the
HTML does not update even though the backend has updated.

* Update organizations to be a public route

* Update naming with error to exception

* Update Functionality in Organization Editor (#104)

* Add edit organization button

* Pull current organization into editor

* Edit org editor button

---------

Co-authored-by: jadekeegan <97476936+jadekeegan@users.noreply.github.com>
Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>
Co-authored-by: Brianna Ta <37271052+briannata@users.noreply.github.com>
Co-authored-by: atoneyd <31745478+atoneyd@users.noreply.github.com>
Co-authored-by: tdianaa <rtoneyd@gmail.com>
Co-authored-by: briannata <brianna.hy.ta@gmail.com>