Create individual file for EventDetails model
This commit creates a separate file for the EventDetails model (which
inherits from the Event model) to avoid issues with circular imports.
The Event Entity and original Event model files are refactored according
to this change.

This commit also adds the import for EventDetails to the __init__.py
file for the models so it is included in the packages and can be
imported more cleanly.