Add Organization RXJS Object (#105)
* Update organization details page to use resolver

This commit refactors the Organization Details page to use a functional
resolver to inject the Organization into the application rather than
using subscriptions.

* Update organization details title resolver

This commit fixes the organiztion details page title resolver to ensure
that it does not make an unnecessary extra HTTP request to get the
organization information. This is done by making title/title resolver a
child so it can inherit the parent route data!

* Add Permission Guard to Organization Editor

* Refactor Organization into the Resolver

* Add Social Media Icon Service (#92)

* Add Social Media Icon Service
* Remove Spec File

* Update the admin organization page to use resolvers

This commit refactors the admin organization page to use route resolvers
for consistency across components. It unfortunately does not auto update
however as the route data does not get refreshed when the database
changes.

* Clean up imports

This commit removes unnecessary imports from the organization details
component and the organization editor component.

* Improve Organization Slugs (#90)

* Add shorthand to organization backend

* Add organization shorthand to frontend

* Add automatic slug generation for org creation

* Add documentation to change method

* Add alembic migration to add shorthand column (#97)

* Add form validation (#91)

* Add form validation

* Add hint specification

* Update Organization Details Page to use short description if no long description

---------

Co-authored-by: Brianna Ta <37271052+briannata@users.noreply.github.com>

* Clean Up Organization Backend (#95)

* Check admin permission for org update

* Make get from slug more idiomatic

* Fix organization create id nit

* Attempt to implement rxjs object for organizations

* Fix the admin organization RXJS object behavior

This commit resolves issues with the admin organization RXJS object so
it can be properly used in the admin list component.

* Remove Organizations interface

This commit removes the Organizations interface and replaces any instance
of Organizations with Organization[] to simplify the code structure.

* Remove extra import from resolving merge conflicts

* Remove Admin Organization resolver

Remove the admin organization resolver from the frontend code.

* Clean up imports and unnecessary code

This commit removes unnecessary code/imports. It also fixes some imports
that unnecessarily use the full path or use a less simplified relative
path for certain files.

---------

Co-authored-by: Ajay Gandecha <ajgandecha@gmail.com>
Co-authored-by: Brianna Ta <37271052+briannata@users.noreply.github.com>
Co-authored-by: Kris Jordan <kris@cs.unc.edu>
Co-authored-by: atoneyd <31745478+atoneyd@users.noreply.github.com>