Launch XL Coworking with Dropins (#109)
The goal of the UNC CS Experience Labs' Coworking facilities is to give students industry-grade workstations, equipment, and collaboration spaces. The XL web app aims to be how students make drop-in and future reservations for these facilities.

This commit launches drop-in capabilities for single seats in SN156 and sets the foundation for more progress to come with future reservations, group reservations, and room reservations.

For students, a new home page view is added with the operating hours of the XL and a listing of currently available seats in SN156. Clicking a seat begins a reservation workflow for a drop-in.

For ambassadors, a new view is added for checking in students who have requested a drop-in via the XL app.