Implement the Frontend for the Events Feature (#108)
* Create Wireframe, Add Start of Dialog

* Separate Events into Separate Pages

This commit separates the events page and the events detail page.

* Remove Experimental Component, Move it into Events

This commit removes the Experimental Component and moves it into the Event Component.

* Link Event Cards to the Detail Page

* Label Event Detail Header with TODO

* Reimplement Column Frontend Wireframe

* Create Events Service and Resolvers

This commit creates the event service and resolvers, as well as created an initial connection between the events backend and frontend wireframe.

* Set Up Event Details Component

This commit sets up the event details component so that it takes in an event ID and returns the correct page.

* Disable Event Card Links when not on Mobile

This commit disables the event card links when the view is not on a mobile device.

* Group Events by Date

This commit groups events by their date in the events page.

* Fix Padding on Events Sidebar

This commit fixes some spacing issues on the events sidebar.

* Fix Description Text and Spacing

This commit fixes some more styling on the events cards and events page.

* Implement Event Search Bar

This commit implements the search bar for the events page.

* Implement Event Share Button

This commit implements the share button by allowing users to copy the event URL to the clipboard.

* Add More Demo Event Data

This commit adds more demo event data to demo the scrolling features.

* Fix Event Card Highlighting When Window Resized

This commit ensures that the highlighting on event cards go away if the sidebar is hidden.

* Add Documentation to the Events Frontend

This commit adds documentation to the events frontend code, namely both component files, the resolvers, and service.

* Fix Highlighting Colors on Light Mode

This commit fixes highlighting color issues when on light mode.

* Fix Spacing Issue

This commit fixes another spacing issue in the events grid.

* Filter out Past Events

This commit ensures that all past events are filtered out of the view.

* Revert Previous Commit

* Fix Pytest Errors

This commit fixes errors in Pytest by fixing residual errors from a reverted commit.

* Fix Pixel Space Left Above Search Bar

This commit fixes the 1px gap remaining at the top of the search bar.

* Change Cursor to Pointer on Events Org Link Hover

This commit changes the cursor to a pointer shape when hovering over organization links on the event detail card.

* Fix Spacing Issue Between Event Tiles

This commit fixes issues with spacing between event tiles.

This issue was actually caused by the appearance of the scroll bar. To fix this, I just hid the scroll bar on all browsers and changed the detail column's right position back to the original 24px that it was.

* Clean Event Details Title

This commit changes the default title of the Event Details page from 'Event Detail - TODO: Make Title Dynamic' to just 'Event Details'. This text is overwritten anyway by the result of the title resolver, but it serves as the default in case that operation fails.

* Remove Description Text from Event Card

This commit removes the description text from the event card.

* Refactor Event Grouping Method

This commit refactors the event grouping method to return a map rather than to reinitialize the eventsPerDay field. In addition, this commit also adds the keyvalue pipe to the frontend HTML to prevent the expression changed after initialization error in Angular.

* Remove Commented Code

This commit removes some commented code from the Event Page leftover from the wireframe stage.

* Remove Trailing Divider

This commit removes the trailing divider in the event details card.